backend/requirements.txt
# HiddyStays Backend Dependencies

# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Payment processing
stripe>=7.0.0

# Database and ORM
supabase>=2.0.0

# HTTP client
httpx>=0.25.0

# Data validation and serialization
python-multipart>=0.0.6

# Logging
structlog>=23.2.0

# Email services
resend>=0.6.0
jinja2>=3.1.0

# Security and authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Date and time handling
python-dateutil>=2.8.2

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.7.0

# Production
gunicorn>=21.2.0
